requirements.txt
luigi==2.8.13
praw==7.0.0
requests==2.23.0
jsonlib-python3==1.6.1
parameterized==0.7.4
configparser==5.0.0
psycopg2-binary==2.8.5